+++ crates/node-common/src/package_json.rs
@@ -6,4 +6,6 @@ pub struct VoltaField {
     pub npm: Option<String>,
     pub pnpm: Option<String>,
     pub yarn: Option<String>,
+    // Special
+    pub extends: Option<String>,
 }