+++ tools/node/src/proto.rs
@@ -161,7 +161,7 @@ pub fn download_prebuilt(
         HostArch::Arm => "armv7l".into(),
         HostArch::Arm64 => "arm64".into(),
         HostArch::Powerpc64 => {
-            if env.os == HostOS::Linux {
+            if env.os.is_linux() {
                 "ppc64le".into()
             } else {
                 "ppc64".into()
@@ -210,7 +210,7 @@ pub fn download_prebuilt(
         _ => unreachable!(),
     };
 
-    let filename = if env.os == HostOS::Windows {
+    let filename = if env.os.is_windows() {
         format!("{prefix}.zip")
     } else {
         format!("{prefix}.tar.xz")