+++ tools/node-depman/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 #### 🚀 Updates
 
+- Added support for `volta.extends`: https://docs.volta.sh/advanced/workspaces
+- Updated `volta` to take precedence over `engines` in `package.json`.
 - Updated to support proto v0.42 release.
 
 ## 0.13.1